+++ Homie Hand/src/assets/Lava.java
@@ -11,9 +11,9 @@ public class Lava extends GameObject {
 	// r = 51, g = 0, b = 0
 	private BufferedImage lava;
 	
-	public Lava(int x, int y, int width, int height, int scale, String ID) {
+	public Lava(float x, float y, float width, float height, float scale, String ID) {
 		super(x, y, width, height, scale, ID);
-		lava = ImageLoader.getSpriteSheet(ImageLoader.BLOCK_SPRITESHEET).getSubimage(width * 3, 0, width, height);
+		lava = ImageLoader.getSpriteSheet(ImageLoader.BLOCK_SPRITESHEET).getSubimage((int)width * 3, 0, (int)width, (int)height);
 	}
 	
 	public void tick(LinkedList<GameObject> gameObjects) {
@@ -21,6 +21,6 @@ public class Lava extends GameObject {
 	}
 	
 	public void render(Graphics g) {
-		g.drawImage(lava, x, y, scaledWidth, scaledHeight, null);
+		g.drawImage(lava, (int)x, (int)y, (int)scaledWidth, (int)scaledHeight, null);
 	}
 }
